\name{ecolUtil-package}
\alias{ecolUtil-package}
\alias{ecolUtil}
\docType{package}
\title{
\packageTitle{ecolUtil}
}
\description{
\packageDescription{ecolUtil}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecolUtil}
\packageIndices{ecolUtil}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ecolUtil}

Maintainer: \packageMaintainer{ecolUtil}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
